streamlit_app/requirements.txt
ipython==8.8.0
ipython-genutils==0.2.0
joblib==1.2.0
keras==2.9.0
matplotlib==3.6.1
pandas==1.3.3
plotly==5.11.0
python-doctr==0.6.0
seaborn==0.12.0
sentencepiece==0.1.97
streamlit==1.12.2
tensorflow==2.9.2
tensorflow-addons==0.18.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.27.0
torch==1.12.0
torchvision==0.13.0
tf2onnx==1.12.1
transformers==4.25.1
